+++ src/Environment/Environment.php
@@ -22,7 +22,7 @@ use League\CommonMark\Configuration\ConfigurationInterface;
 use League\CommonMark\Delimiter\DelimiterParser;
 use League\CommonMark\Delimiter\Processor\DelimiterProcessorCollection;
 use League\CommonMark\Delimiter\Processor\DelimiterProcessorInterface;
-use League\CommonMark\Event\DocumentRenderedEvent;
+use League\CommonMark\Event\DocumentParsedEvent;
 use League\CommonMark\Event\ListenerData;
 use League\CommonMark\Extension\CommonMark\CommonMarkCoreExtension;
 use League\CommonMark\Extension\ConfigurableExtensionInterface;
@@ -31,6 +31,7 @@ use League\CommonMark\Extension\GithubFlavoredMarkdownExtension;
 use League\CommonMark\Normalizer\SlugNormalizer;
 use League\CommonMark\Normalizer\TextNormalizerInterface;
 use League\CommonMark\Normalizer\UniqueSlugNormalizer;
+use League\CommonMark\Normalizer\UniqueSlugNormalizerInterface;
 use League\CommonMark\Parser\Block\BlockStartParserInterface;
 use League\CommonMark\Parser\Inline\InlineParserInterface;
 use League\CommonMark\Renderer\NodeRendererInterface;
@@ -408,8 +409,17 @@ final class Environment implements EnvironmentInterface, EnvironmentBuilderInter
             \assert($normalizer instanceof TextNormalizerInterface);
             $this->injectEnvironmentAndConfigurationIfNeeded($normalizer);
 
-            $this->slugNormalizer = new UniqueSlugNormalizer($normalizer, $this->config->get('slug_normalizer/scope'));
-            $this->addEventListener(DocumentRenderedEvent::class, [$this->slugNormalizer, 'onDocumentRendered'], -1000);
+            if ($this->config->get('slug_normalizer/unique') !== UniqueSlugNormalizerInterface::DISABLED && ! $normalizer instanceof UniqueSlugNormalizer) {
+                $normalizer = new UniqueSlugNormalizer($normalizer);
+            }
+
+            if ($normalizer instanceof UniqueSlugNormalizer) {
+                if ($this->config->get('slug_normalizer/unique') === UniqueSlugNormalizerInterface::PER_DOCUMENT) {
+                    $this->addEventListener(DocumentParsedEvent::class, [$normalizer, 'clearHistory'], -1000);
+                }
+            }
+
+            $this->slugNormalizer = $normalizer;
         }
 
         return $this->slugNormalizer;
@@ -439,7 +449,7 @@ final class Environment implements EnvironmentInterface, EnvironmentBuilderInter
             'slug_normalizer' => Expect::structure([
                 'instance' => Expect::type(TextNormalizerInterface::class)->default(new SlugNormalizer()),
                 'max_length' => Expect::int()->min(0)->default(255),
-                'scope' => Expect::anyOf(UniqueSlugNormalizer::SCOPE_ENVIRONMENT, UniqueSlugNormalizer::SCOPE_DOCUMENT)->default(UniqueSlugNormalizer::SCOPE_DOCUMENT),
+                'unique' => Expect::anyOf(UniqueSlugNormalizerInterface::DISABLED, UniqueSlugNormalizerInterface::PER_ENVIRONMENT, UniqueSlugNormalizerInterface::PER_DOCUMENT)->default(UniqueSlugNormalizerInterface::PER_DOCUMENT),
             ])->castTo('array'),
         ]);
     }