+++ tests/unit/Environment/EnvironmentTest.php
@@ -23,8 +23,10 @@ use League\CommonMark\Delimiter\Processor\DelimiterProcessorInterface;
 use League\CommonMark\Environment\Environment;
 use League\CommonMark\Environment\EnvironmentBuilderInterface;
 use League\CommonMark\Event\AbstractEvent;
+use League\CommonMark\Event\DocumentParsedEvent;
 use League\CommonMark\Extension\ConfigurableExtensionInterface;
 use League\CommonMark\Extension\ExtensionInterface;
+use League\CommonMark\Node\Block\Document;
 use League\CommonMark\Normalizer\TextNormalizerInterface;
 use League\CommonMark\Parser\Block\BlockStartParserInterface;
 use League\CommonMark\Parser\Inline\InlineParserInterface;
@@ -473,15 +475,8 @@ class EnvironmentTest extends TestCase
 
     public function testCustomSlugNormalizer(): void
     {
-        $innerNormalizer = new class implements TextNormalizerInterface {
-            /**
-             * {@inheritDoc}
-             */
-            public function normalize(string $text, array $context = []): string
-            {
-                return 'foo';
-            }
-        };
+        $innerNormalizer = $this->createStub(TextNormalizerInterface::class);
+        $innerNormalizer->method('normalize')->willReturn('foo');
 
         $environment = new Environment([
             'slug_normalizer' => [
@@ -490,8 +485,62 @@ class EnvironmentTest extends TestCase
         ]);
 
         $normalizer = $environment->getSlugNormalizer();
-        $this->assertSame('foo', $normalizer->normalize('Test'));
-        $this->assertSame('foo-1', $normalizer->normalize('Something else'));
+        $this->assertSame('foo', $normalizer->normalize('Foo'));
+        $this->assertSame('foo-1', $normalizer->normalize('Foo'));
+    }
+
+    public function testUniqueSlugNormalizerDisabled(): void
+    {
+        $environment = new Environment([
+            'slug_normalizer' => [
+                'unique' => false,
+            ],
+        ]);
+
+        $normalizer = $environment->getSlugNormalizer();
+        $this->assertSame('foo', $normalizer->normalize('Foo'));
+        $this->assertSame('foo', $normalizer->normalize('Foo'));
+        $this->assertSame('foo', $normalizer->normalize('Foo'));
+    }
+
+    public function testUniqueSlugNormalizerPerDocument(): void
+    {
+        $environment = new Environment([
+            'slug_normalizer' => [
+                'unique' => 'document',
+            ],
+        ]);
+
+        $normalizer = $environment->getSlugNormalizer();
+        $this->assertSame('foo', $normalizer->normalize('Foo'));
+        $this->assertSame('foo-1', $normalizer->normalize('Foo'));
+        $this->assertSame('foo-2', $normalizer->normalize('Foo'));
+
+        $environment->dispatch(new DocumentParsedEvent(new Document()));
+
+        $this->assertSame('foo', $normalizer->normalize('Foo'));
+        $this->assertSame('foo-1', $normalizer->normalize('Foo'));
+        $this->assertSame('foo-2', $normalizer->normalize('Foo'));
+    }
+
+    public function testUniqueSlugNormalizerPerEnvironment(): void
+    {
+        $environment = new Environment([
+            'slug_normalizer' => [
+                'unique' => 'environment',
+            ],
+        ]);
+
+        $normalizer = $environment->getSlugNormalizer();
+        $this->assertSame('foo', $normalizer->normalize('Foo'));
+        $this->assertSame('foo-1', $normalizer->normalize('Foo'));
+        $this->assertSame('foo-2', $normalizer->normalize('Foo'));
+
+        $environment->dispatch(new DocumentParsedEvent(new Document()));
+
+        $this->assertSame('foo-3', $normalizer->normalize('Foo'));
+        $this->assertSame('foo-4', $normalizer->normalize('Foo'));
+        $this->assertSame('foo-5', $normalizer->normalize('Foo'));
     }
 
     /**