+++ CHANGELOG.md
@@ -17,7 +17,7 @@ See <https://commonmark.thephpleague.com/2.0/upgrading/> for detailed informatio
  - Added `heading_permalink/min_heading_level` and `heading_permalink/max_heading_level` options to control which headings get permalinks (#519)
  - Added `footnote/backref_symbol` option for customizing backreference link appearance (#522)
  - Added `slug_normalizer/max_length` option to control the maximum length of generated URL slugs
- - Added `slug_normalizer/scope` option to control whether unique slugs should be generated per-document or per-environment
+ - Added `slug_normalizer/unique` option to control whether unique slugs should be generated per-document or per-environment
  - Added new `HtmlFilter` and `StringContainerHelper` utility classes
  - Added new `AbstractBlockContinueParser` class to simplify the creation of custom block parsers
  - Added several new classes and interfaces:
@@ -50,6 +50,7 @@ See <https://commonmark.thephpleague.com/2.0/upgrading/> for detailed informatio
    - `RenderedContentInterface`
    - `SpecReader`
    - `UniqueSlugNormalizer`
+   - `UniqueSlugNormalizerInterface`
  - Added several new methods:
    - `ConfigurationInterface::exists()`
    - `Environment::createDefaultConfiguration()`